Django/website1/requirements.txt
django
stripe
bcrypt==4.2.0
django-filter==24.3
djangorestframework
pillow
python-dotenv